+++ src/main/java/ro/webdata/normalization/timespan/ro/regex/TimespanRegex.java
@@ -110,7 +110,7 @@ public class TimespanRegex {
     public static final String START_END = END + START;
     public static final String CENTURY_LABEL = "(" + START_END +  "\\s*(?:(secol|secoi)(?:ele|ului|ul)?|sec)[\\.\\s]*" + ARTICLE_AL + ")";
     public static final String MILLENNIUM_LABEL = "(" + START_END + "\\s*(?:mileni(?:ile|ului|ul)?|mil)[\\.\\s]*" + ARTICLE_AL + ")";
-    public static final String YEAR_LABEL = "(" + START_END + "\\s*(?:ani+|an(?:ului|ulu|ul|u)?|an)[\\.\\s]*" + ARTICLE_AL + ")";
+    public static final String YEAR_LABEL = "(" + START_END + "\\s*(?:ani+(?:lor)?|an(?:ului|ulu|ul|u)?)\\s*" + ARTICLE_AL + ")";
 
     private static final String HALF = "jum([aă]tate(a)?)?";
     private static final String FIRST_HALF_STRING_REGEX = "(" + "prim[a]*[\\. ]+(" + HALF + "|part)" + ")";